+++ src/Command/ConfigureContainerCommand.php
@@ -1,13 +1,11 @@
 <?php
 namespace Oka\FileBundle\Command;
 
-use Oka\FileBundle\Model\File;
 use Oka\FileBundle\Utils\FileUtil;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Filesystem\Filesystem;
 
 /**
  *